+++ HISTORY.rst
@@ -5,6 +5,12 @@ picows Release History
    :depth: 1
    :local:
 
+2.1.0 ()
+
+* Websockets: add open, closed attributes to ConnectionBase for compatibility with legacy API.
+* Websockets: add protocol module with State enum for compatibility
+* Websockets: alias WebsocketClientProtocol to ClientConnection for compatibility with legacy API
+
 2.0.0 (2026-05-18)
 -----------------